+++ src/main/java/com/hedvig/notificationService/service/event/EventRequest.kt
@@ -1,5 +1,4 @@
-// TODO move this package with the event handler
-package com.hedvig.notificationService.customerio.dto
+package com.hedvig.notificationService.service.event
 
 import com.fasterxml.jackson.annotation.JsonSubTypes
 import com.fasterxml.jackson.annotation.JsonTypeInfo
