+++ src/test/java/com/hedvig/notificationService/customerio/customerioEvents/jobs/JobSchedulerTest.kt
@@ -3,7 +3,7 @@ package com.hedvig.notificationService.customerio.customerioEvents.jobs
 import assertk.assertThat
 import assertk.assertions.isEqualTo
 import com.hedvig.notificationService.customerio.SIGN_EVENT_WINDOWS_SIZE_MINUTES
-import com.hedvig.notificationService.customerio.dto.StartDateUpdatedEvent
+import com.hedvig.notificationService.service.event.StartDateUpdatedEvent
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.slot
@@ -37,7 +37,11 @@ class JobSchedulerTest {
         every { scheduler.rescheduleJob(any(), any()) } returns Date()
 
         jobScheduler.rescheduleOrTriggerStartDateUpdated(
-            StartDateUpdatedEvent("contractId", "someMemberId", LocalDate.of(2020, 10, 1)),
+            StartDateUpdatedEvent(
+                "contractId",
+                "someMemberId",
+                LocalDate.of(2020, 10, 1)
+            ),
             callTime
         )
