+++ src/test/java/com/hedvig/notificationService/customerio/web/ContractCreatedEventCreatesContractTest.kt
@@ -8,7 +8,7 @@ import assertk.assertions.isNotNull
 import com.hedvig.notificationService.customerio.ConfigurationProperties
 import com.hedvig.notificationService.customerio.CustomerioService
 import com.hedvig.notificationService.service.event.EventHandler
-import com.hedvig.notificationService.customerio.dto.ContractCreatedEvent
+import com.hedvig.notificationService.service.event.ContractCreatedEvent
 import com.hedvig.notificationService.customerio.hedvigfacades.MemberServiceImpl
 import com.hedvig.notificationService.customerio.state.CustomerioState
 import com.hedvig.notificationService.customerio.state.InMemoryCustomerIOStateRepository
@@ -51,7 +51,11 @@ class ContractCreatedEventCreatesContractTest {
     internal fun `after contract created with no existing contracts a contract exists`() {
 
         sut.onContractCreatedEventHandleRequest(
-            ContractCreatedEvent("theContractId", "aMemberId", null),
+            ContractCreatedEvent(
+                "theContractId",
+                "aMemberId",
+                null
+            ),
             Instant.parse("2020-06-03T08:02:39.403803Z")
         )
 
@@ -74,7 +78,11 @@ class ContractCreatedEventCreatesContractTest {
         repo.save(customerioState)
 
         sut.onContractCreatedEventHandleRequest(
-            ContractCreatedEvent("theNewContractId", "aMemberId", null),
+            ContractCreatedEvent(
+                "theNewContractId",
+                "aMemberId",
+                null
+            ),
             Instant.parse("2020-06-03T08:02:39.403803Z")
         )
 
@@ -97,7 +105,11 @@ class ContractCreatedEventCreatesContractTest {
         repo.save(customerioState)
 
         sut.onContractCreatedEventHandleRequest(
-            ContractCreatedEvent("theFirstContractId", "aMemberId", null),
+            ContractCreatedEvent(
+                "theFirstContractId",
+                "aMemberId",
+                null
+            ),
             Instant.parse("2020-06-03T08:02:39.403803Z")
         )
