+++ src/test/java/com/hedvig/notificationService/customerio/web/OnContractCreatedEventTest.kt
@@ -11,7 +11,7 @@ import com.hedvig.notificationService.customerio.CustomerioService
 import com.hedvig.notificationService.service.event.EventHandler
 import com.hedvig.notificationService.customerio.SIGN_EVENT_WINDOWS_SIZE_MINUTES
 import com.hedvig.notificationService.customerio.customerioEvents.jobs.ContractCreatedJob
-import com.hedvig.notificationService.customerio.dto.ContractCreatedEvent
+import com.hedvig.notificationService.service.event.ContractCreatedEvent
 import com.hedvig.notificationService.customerio.hedvigfacades.ContractLoader
 import com.hedvig.notificationService.customerio.hedvigfacades.MemberServiceImpl
 import com.hedvig.notificationService.customerio.state.CustomerioState
