+++ src/test/java/com/hedvig/notificationService/customerio/web/OnContractRenewalQuuedEventTest.kt
@@ -4,7 +4,7 @@ import assertk.assertThat
 import assertk.assertions.isEqualTo
 import com.hedvig.notificationService.customerio.CustomerioService
 import com.hedvig.notificationService.service.event.EventHandler
-import com.hedvig.notificationService.customerio.dto.ContractRenewalQueuedEvent
+import com.hedvig.notificationService.service.event.ContractRenewalQueuedEvent
 import com.hedvig.notificationService.customerio.hedvigfacades.MemberServiceImpl
 import com.hedvig.notificationService.customerio.state.InMemoryCustomerIOStateRepository
 import com.hedvig.notificationService.service.firebase.FirebaseNotificationService
